+++ common/exception/src/exception.rs
@@ -30,10 +30,10 @@ macro_rules! build_exceptions {
         as_item! {
             impl ErrorCodes {
                 $(
-                pub fn $body(display_text: String) -> ErrorCodes {
+                pub fn $body(display_text: impl Into<String>) -> ErrorCodes {
                     ErrorCodes {
                         code:$code,
-                        display_text,
+                        display_text: display_text.into(),
                         cause: None,
                         backtrace: Some(Backtrace::new()),
                     }
@@ -70,6 +70,7 @@ build_exceptions! {
     CannotReadFile(23),
     ParquetError(24),
     UnknownTable(25),
+    IllegalAggregateExp(26),
 
     UnknownException(1000),
     TokioError(1001)