+++ fusequery/query/src/pipelines/processors/processor_merge.rs
@@ -14,14 +14,19 @@ use tokio_stream::wrappers::ReceiverStream;
 use tokio_stream::StreamExt;
 
 use crate::pipelines::processors::IProcessor;
+use crate::sessions::FuseQueryContextRef;
 
 pub struct MergeProcessor {
+    ctx: FuseQueryContextRef,
     inputs: Vec<Arc<dyn IProcessor>>
 }
 
 impl MergeProcessor {
-    pub fn create() -> Self {
-        MergeProcessor { inputs: vec![] }
+    pub fn create(ctx: FuseQueryContextRef) -> Self {
+        MergeProcessor {
+            ctx,
+            inputs: vec![]
+        }
     }
 }
 
@@ -48,7 +53,7 @@ impl IProcessor for MergeProcessor {
         let inputs = self.inputs.len();
         match inputs {
             0 => Result::Err(ErrorCodes::IllegalTransformConnectionState(
-                "Merge processor inputs cannot be zero".to_string()
+                "Merge processor inputs cannot be zero"
             )),
             1 => self.inputs[0].execute().await,
             _ => {
@@ -56,7 +61,7 @@ impl IProcessor for MergeProcessor {
                 for i in 0..inputs {
                     let input = self.inputs[i].clone();
                     let sender = sender.clone();
-                    tokio::spawn(async move {
+                    self.ctx.execute_task(async move {
                         let mut stream = match input.execute().await {
                             Err(e) => {
                                 sender.send(Result::Err(e)).await.ok();