+++ fusestore/store/src/meta_service/meta.rs
@@ -2,22 +2,26 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
-use std::collections::hash_map::DefaultHasher;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
-use std::hash::Hash;
-use std::hash::Hasher;
+use std::fmt;
+use std::fmt::Display;
+use std::fmt::Formatter;
 
-use async_trait::async_trait;
+use serde::Deserialize;
+use serde::Serialize;
 
-#[async_trait]
-pub trait IMeta {
-    async fn add(&mut self, key: String, value: String) -> anyhow::Result<()>;
-    // async fn remove(&self, key: String) -> anyhow::Result<()>;
-}
+use crate::meta_service::ClientRequest;
+use crate::meta_service::ClientResponse;
+use crate::meta_service::Cmd;
+use crate::meta_service::IPlacement;
+use crate::meta_service::NodeId;
 
-/// Distributed FS meta data manager
-#[derive(Default)]
+/// Meta data of a Dfs.
+/// Includes:
+/// - what files are stored in this Dfs.
+/// - the algo about how to distribute files to nodes.
+#[derive(Serialize, Deserialize, Debug, Default, Clone)]
 pub struct Meta {
     /// The file names stored in this cluster
     pub keys: BTreeMap<String, String>,
@@ -27,53 +31,77 @@ pub struct Meta {
     pub nodes: HashMap<NodeId, Node>
 }
 
+impl Meta {
+    // Apply an op from client.
+    pub fn apply(&mut self, data: &ClientRequest) -> anyhow::Result<ClientResponse> {
+        match data.cmd {
+            Cmd::AddFile { ref key, ref value } => {
+                if self.keys.contains_key(key) {
+                    let prev = self.keys.get(key);
+                    Ok((prev.cloned(), None).into())
+                } else {
+                    let prev = self.keys.insert(key.clone(), value.clone());
+                    Ok((prev, Some(value.clone())).into())
+                }
+            }
+
+            Cmd::SetFile { ref key, ref value } => {
+                let prev = self.keys.insert(key.clone(), value.clone());
+                Ok((prev, Some(value.clone())).into())
+            }
+
+            Cmd::AddNode {
+                ref node_id,
+                ref node
+            } => {
+                if self.nodes.contains_key(node_id) {
+                    let prev = self.nodes.get(node_id);
+                    Ok((prev.cloned(), None).into())
+                } else {
+                    let prev = self.nodes.insert(*node_id, node.clone());
+                    Ok((prev, Some(node.clone())).into())
+                }
+            }
+        }
+    }
+
+    pub fn get_file(&self, key: &str) -> Option<String> {
+        let x = self.keys.get(key);
+        x.cloned()
+    }
+
+    pub fn get_node(&self, node_id: &NodeId) -> Option<Node> {
+        let x = self.nodes.get(node_id);
+        x.cloned()
+    }
+}
+
 /// A slot is a virtual and intermediate allocation unit in a distributed storage.
 /// The key of an object is mapped to a slot by some hashing algo.
 /// A slot is assigned to several physical servers(normally 3 for durability).
+#[derive(Serialize, Deserialize, Debug, Default, Clone)]
 pub struct Slot {
-    node_ids: Vec<i64>
+    pub node_ids: Vec<NodeId>
 }
 
-pub type NodeId = i64;
-
-#[derive(Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq)]
 pub struct Node {
     pub name: String,
     pub address: String
 }
 
-#[async_trait]
-impl IMeta for Meta {
-    async fn add(&mut self, key: String, value: String) -> anyhow::Result<()> {
-        self.keys.insert(key, value);
-        Ok(())
+impl Display for Node {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(f, "{}={}", self.name, self.address)
     }
 }
 
-impl Meta {
-    pub fn empty() -> Self {
-        Self {
-            ..Default::default()
-        }
+impl IPlacement for Meta {
+    fn get_slots(&self) -> &[Slot] {
+        &self.slots
     }
 
-    /// Returns the slot index to store a file.
-    pub fn slot_index_for_key(&self, key: &str) -> u64 {
-        // TODO use consistent hash if need to extend cluster.
-        let mut hasher = DefaultHasher::new();
-        key.hash(&mut hasher);
-        let hsh = hasher.finish();
-        hsh % self.slots.len() as u64
-    }
-
-    /// Returns the Node-s that keeps a copy of a file.
-    pub fn nodes_to_store_key(&self, key: &str) -> Vec<Node> {
-        let slot_idx = self.slot_index_for_key(key);
-        let slot = &self.slots[slot_idx as usize];
-
-        slot.node_ids
-            .iter()
-            .map(|nid| (*self.nodes.get(nid).unwrap()).clone())
-            .collect()
+    fn get_node(&self, node_id: &u64) -> Option<&Node> {
+        self.nodes.get(node_id)
     }
 }