+++ fusestore/store/src/meta_service/meta_service_impl_test.rs
@@ -6,12 +6,14 @@
 use log::info;
 use pretty_assertions::assert_eq;
 
+use crate::meta_service::ClientRequest;
+use crate::meta_service::ClientResponse;
+use crate::meta_service::Cmd;
 use crate::meta_service::GetReq;
 use crate::meta_service::MetaNode;
 use crate::meta_service::MetaServiceClient;
 use crate::meta_service::MetaServiceImpl;
 use crate::meta_service::MetaServiceServer;
-use crate::meta_service::SetReq;
 use crate::tests::rand_local_addr;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
@@ -19,6 +21,8 @@ async fn test_meta_server_set_get() -> anyhow::Result<()> {
     let addr = rand_local_addr();
 
     let mn = MetaNode::new(0).await;
+    let rst = mn.boot(addr.clone()).await;
+    assert!(rst.is_ok());
 
     let meta_srv_impl = MetaServiceImpl::create(mn).await;
     let meta_srv = MetaServiceServer::new(meta_srv_impl);
@@ -29,13 +33,23 @@ async fn test_meta_server_set_get() -> anyhow::Result<()> {
 
     {
         // add: ok
-        let req = tonic::Request::new(SetReq {
-            key: "foo".into(),
-            value: "bar".into(),
-            if_absent: true
-        });
-        let rst = client.set(req).await?.into_inner();
-        assert_eq!(true, rst.ok);
+        let req = ClientRequest {
+            txid: None,
+            cmd: Cmd::AddFile {
+                key: "foo".to_string(),
+                value: "bar".to_string()
+            }
+        };
+        let rst = client.write(req).await?.into_inner();
+        let resp: ClientResponse = rst.into();
+        match resp {
+            ClientResponse::String { prev: _, result } => {
+                assert_eq!("bar".to_string(), result.unwrap());
+            }
+            _ => {
+                panic!("not string")
+            }
+        }
 
         // get the stored value
 
@@ -46,23 +60,43 @@ async fn test_meta_server_set_get() -> anyhow::Result<()> {
 
     {
         // add: conflict with existent.
-        let req = tonic::Request::new(SetReq {
-            key: "foo".into(),
-            value: "bar".into(),
-            if_absent: true
-        });
-        let rst = client.set(req).await?.into_inner();
-        assert_eq!(false, rst.ok);
+        let req = ClientRequest {
+            txid: None,
+            cmd: Cmd::AddFile {
+                key: "foo".to_string(),
+                value: "bar".to_string()
+            }
+        };
+        let rst = client.write(req).await?.into_inner();
+        let resp: ClientResponse = rst.into();
+        match resp {
+            ClientResponse::String { prev: _, result } => {
+                assert!(result.is_none());
+            }
+            _ => {
+                panic!("not string")
+            }
+        }
     }
     {
         // set: overrde. ok.
-        let req = tonic::Request::new(SetReq {
-            key: "foo".into(),
-            value: "bar2".into(),
-            if_absent: false
-        });
-        let rst = client.set(req).await?.into_inner();
-        assert_eq!(true, rst.ok);
+        let req = ClientRequest {
+            txid: None,
+            cmd: Cmd::SetFile {
+                key: "foo".to_string(),
+                value: "bar2".to_string()
+            }
+        };
+        let rst = client.write(req).await?.into_inner();
+        let resp: ClientResponse = rst.into();
+        match resp {
+            ClientResponse::String { prev: _, result } => {
+                assert_eq!(Some("bar2".to_string()), result);
+            }
+            _ => {
+                panic!("not string")
+            }
+        }
 
         // get the stored value