+++ fusestore/store/src/meta_service/mod.rs
@@ -4,14 +4,15 @@
 
 mod meta;
 mod meta_service_impl;
+mod placement;
 mod raftmeta;
 
-pub use meta::IMeta;
+pub use async_raft::NodeId;
 pub use meta::Meta;
 pub use meta::Node;
-pub use meta::NodeId;
 pub use meta::Slot;
 pub use meta_service_impl::MetaServiceImpl;
+pub use placement::IPlacement;
 pub use raftmeta::ClientRequest;
 pub use raftmeta::ClientResponse;
 pub use raftmeta::Cmd;
@@ -28,8 +29,6 @@ pub use crate::protobuf::meta_service_server::MetaServiceServer;
 pub use crate::protobuf::GetReply;
 pub use crate::protobuf::GetReq;
 pub use crate::protobuf::RaftMes;
-pub use crate::protobuf::SetReply;
-pub use crate::protobuf::SetReq;
 
 #[cfg(test)]
 mod meta_service_impl_test;