+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] 2021-05-13
+
+### Added
+* [FuseQuery] Support ClickHouse query progress - Thanks @sundy-li
+* [FuseQuery] Support Error Code - Thanks @zhang2014
+* [FuseQuery] Support `substring` function ([#487](https://github.com/datafuselabs/datafuse/pull/487)) - Thanks @tlightsky
 
 ## [0.4.0] 2021-04-16